screening: add pydantic code and run more tests
this works in general, but llama fails to process
this noisy context - something must be done
around that.

also, everything must be aggressively persisted
from pubmed requests to schemas, yet schemas
actually persist already thanks to find_latest logic.

It's expected that generation is slow because
num_batches is set to 1.